rename: Store 파일들에 접두사로 use 추가
