+++ src/stores/useErrorStore.ts
