+++ src/stores/useNewsletterStore.ts
